feat: nestedBaseViewModel inherits its methods to each viewModel
- change viewModel member variable name in fragment collect
