+++ app/src/main/java/com/example/calefit/ui/home/detali/ExerciseDetailFragment.kt
@@ -94,7 +94,7 @@ class ExerciseDetailFragment : Fragment() {
 
     private fun observeData() {
         viewLifecycleOwner.repeatOnLifecycleExtension {
-            viewModel.exerciseList.collect { exerciseList ->
+            viewModel.exercisePlan.collect { exerciseList ->
                 detailAdapter.submitList(exerciseList.list)
                 binding.templateTitle = exerciseList.templateName
             }
